Refactor Popover stories: args forwarding, utils extraction, z-index fix
- Forward all args to Popover.Root in every render-based story
  (OverlayPlacement, CrossIframe, CrossIframeWithSlotFill) so
  controls are wired to the live component
- Disable children control globally in meta (re-enable for Default
  only), since children JSX is not meaningfully editable via controls
- Extract useMeasure and GenericIframe to stories/utils.tsx with
  JSDoc comments
- Fix WithCustomZIndex to actually set --wp-ui-popover-z-index via
  a scoped style tag and className on the positioner

Made-with: Cursor
